Support for Cordova 8
Seems that the current structure were almost compatible with Cordova 8.
The exception was the usage of the wrong tag for the job of listing a
dependency. Using `<source-file>` to "copy" the `aar` dependency to
the path desired was kind of a trick that worked until now. Cordova 8
gets confused when we do that specially when adding a `libs` folder in
the wrong path.